FastAPI Assignment 2/requirements.txt
nyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
certifi==2023.5.7
click==8.1.4
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.100.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pydantic==1.10.11
PyJWT==2.7.0
pytest==7.4.0
python-decouple==3.8
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
uvicorn==0.22.0